+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This package follows standard semvar, `<major>.<minor>.<build>`. No breaking cha
 * Route navigation is contained fully in a query string inside the hash, it will now no longer affect hash based routing
 * Adds the `nonce` to oauth requests to generate an access token.
 * Add scopes display for security schemes.
+* Fixed @request event `headers` to be a Fetch `headers` object.
 
 ## 0.5 ##
 * Remove deprecated password oauth2 flow